crowd-counting/requirements.txt
gradio==5.50.0
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=9.5.0
scipy>=1.10.0
torch>=2.0.0
torchvision>=0.15.0
git+https://github.com/facebookresearch/segment-anything.git
SAM2>=1.0